adding several functions to the entitys for easier use

+++ src/likes/like.repository.ts
@@ -1,6 +1,7 @@
 import { User } from 'src/auth/user.entity';
 import { EntityRepository, Repository } from 'typeorm';
 import { Post } from '../posts/post.entity';
+import { Comment } from '../comments/comment.entity';
 import { Like } from './like.entity';
 
 @EntityRepository(Like)
@@ -9,26 +10,32 @@ export class LikeRepository extends Repository<Like> {
     super();
   }
 
-  async createLike(post: Post, user: User): Promise<Like> {
+  async createLike(
+    target: Post | Comment,
+    user: User,
+    type: 'post' | 'comment',
+  ): Promise<Like> {
     const like = new Like();
     like.user = user;
-    like.post = post;
+
+    if (type === 'post' && target instanceof Post) {
+      like.post = target;
+    } else if (type === 'comment' && target instanceof Comment) {
+      like.comment = target;
+    } else {
+      throw new Error(`Invalid target type: ${type}`);
+    }
 
     await like.save();
-
-    delete like.user;
-    delete like.post;
-
     return like;
   }
 
-  async deleteLike(post: Post, user: User): Promise<void> {
+  async deleteLike(target: Post | Comment, user: User): Promise<void> {
+    let entityType = target instanceof Post ? 'postId' : 'commentId';
+
     await this.manager.query(
-      `
-      DELETE FROM "like"
-      WHERE "postId" = $1 AND "userId" = $2
-    `,
-      [post.id, user.id],
+      `DELETE FROM "like" WHERE "${entityType}" = $1 AND "userId" = $2`,
+      [target.id, user.id],
     );
   }
 }
